Split to multiple smaller components
